#193 Criando uma única classe para os métodos de acesso ao servidor de Announcements
