+++ hortum_mobile/lib/views/home_customer/services/home_customer_services.dart
@@ -1,9 +1,9 @@
-import 'package:hortum_mobile/data/announ_data_backend.dart';
+import 'package:hortum_mobile/data/announcements/announcements_backend.dart';
 import 'package:hortum_mobile/data/productors_data_backend.dart';
 
 class HomeCustomerServices {
   static populateData(bool isAnnouncements, String filter,
-      AnnounDataApi announData, ProductorsDataApi productorsData) async {
+      AnnouncementsApi announData, ProductorsDataApi productorsData) async {
     if (isAnnouncements) {
       await announData.getAnnoun(filter);
     } else {
